- fix template filename - refactoring parse logic - ifx template
