+++ datamodel_code_generator/main.py
@@ -2,4 +2,4 @@ from datamodel_code_generator.model import BaseModel, DataModelField
 from datamodel_code_generator.parser.openapi import Parser
 
 if __name__ == '__main__':
-    Parser(BaseModel, DataModelField).parse()
+    print(Parser(BaseModel, DataModelField).parse())
