+++ datamodel_code_generator/model/__init__.py
@@ -1,4 +1,4 @@
 from .base import DataModel, DataModelField
-from .custom_roo_type import CustomRootType
+from .custom_root_type import CustomRootType
 from .dataclass import DataClass
 from .pydantic import BaseModel
