+++ datamodel_code_generator/model/custom_root_type.py
