+++ datamodel_code_generator/model/template/pydantic/BaseModel_root.jinja2
@@ -1,5 +1,5 @@
-{% for decorator in decorators -%}
+{%- for decorator in decorators -%}
 {{ decorator }}
-{% endfor -%}
+{%- endfor -%}
 class {{ class_name }}(BaseModel):
     __root__: {{ _type }}
