+++ datamodel_code_generator/model/template/pydantic/dataclass.jinja2
@@ -1,12 +1,12 @@
-{% for decorator in decorators -%}
+{%- for decorator in decorators -%}
 {{ decorator }}
-{% endfor -%}
+{%- endfor -%}
 @dataclass
 class {{ class_name }}:
 {%- for field in fields -%}
-    {% if field.required -%}
+    {%- if field.required %}
     {{ field.name }}: {{ field.type_hint }}
-    {% else -%}
+    {%- else %}
     {{ field.name }}: {{ field.type_hint }} = {{field.default}}
-    {% endif -%}
-{% endfor -%}
+    {%- endif %}
+{%- endfor -%}
