+++ datamodel_code_generator/parser/openapi.py
@@ -1,9 +1,10 @@
 from dataclasses import Field, dataclass
-from typing import Dict, List, Optional, Set, Type
+from typing import Dict, List, Optional, Set, Type, Iterator, Union
 
 from prance import BaseParser, ResolvingParser
 
 from datamodel_code_generator.model import CustomRootType, DataModel, DataModelField
+from datamodel_code_generator.model.base import TemplateBase
 
 
 @dataclass
@@ -35,7 +36,7 @@ data_types: Dict[str, Dict[str, DataType]] = {
 }
 
 
-def get_data_type(_type, format=None) -> DataType:
+def get_data_type(_type, format =None) -> DataType:
     _format: str = format or 'default'
     return data_types[_type][_format]
 
@@ -44,13 +45,19 @@ resolving_parser = ResolvingParser('api.yaml', backend='openapi-spec-validator')
 base_parser = BaseParser('api.yaml', backend='openapi-spec-validator')
 
 
+def dump_templates(templates: Union[TemplateBase, List[TemplateBase]]) -> str:
+    if isinstance(templates, TemplateBase):
+        templates = [templates]
+    return '\n\n'.join(str(m) for m in templates)
+
+
 class Parser:
     def __init__(self, data_model_type: Type[DataModel], data_model_field_type: Type[DataModelField]):
         self.data_model_type: Type[DataModel] = data_model_type
         self.data_model_field_type: Type[DataModelField] = data_model_field_type
         self.models = []
 
-    def parse_object(self, name: str, obj: Dict):
+    def parse_object(self, name: str, obj: Dict) -> str:
         requires: Set[str] = set(obj.get('required', []))
         d_list: List[DataModelField] = []
         for field_name, filed in obj['properties'].items():
@@ -59,24 +66,25 @@ class Parser:
                 name=field_name, type_hint=get_data_type(filed["type"],
                                                          filed.get("format")).type_hint,
                 required=field_name in requires))
-        self.models.append(self.data_model_type(name, fields=d_list))
+        return dump_templates(self.data_model_type(name, fields=d_list))
 
-    def parse_array(self, name: str, obj: Dict):
+    def parse_array(self, name: str, obj: Dict) -> str:
+        templates: List[TemplateBase] = []
         # continue
         if '$ref' in obj['items']:
             _type: str = f"List[{obj['items']['$ref'].split('/')[-1]}]"
-            self.models.append(CustomRootType(name, _type))
+            templates.append(CustomRootType(name, _type))
         elif 'properties' in obj['items']:
             self.parse_object(name[:-1], obj['items'])
-            self.models.append(CustomRootType(name, f'List[{name[:-1]}]'))
+            templates.append(CustomRootType(name, f'List[{name[:-1]}]'))
+        return dump_templates(templates)
 
-    def parse(self):
+    def parse(self) -> str:
+        parsed_objects: List[str] = []
         for obj_name, obj in base_parser.specification['components']['schemas'].items():
             if 'properties' in obj:
-                self.parse_object(obj_name, obj)
+                parsed_objects.append(self.parse_object(obj_name, obj))
             elif 'items' in obj:
-                self.parse_array(obj_name, obj)
+                parsed_objects.append(self.parse_array(obj_name, obj))
 
-        for data_model in self.models:
-            print(data_model)
-            print('')
+        return '\n\n\n'.join(parsed_objects)
